pyfastapi/requirements.txt
fastapi
numpy
uvicorn
geocoder
pytz
timezonefinder
geopy
python-dateutil
pyswisseph
pyjhora
